Rejigger rate limiting in postman a little
